feat(otp-login):handel category deletion assign all posts as uncategorized and delete the category

+++ src/main/java/com/ql/BlogApplication/service/CategoryService.java
@@ -1,8 +1,11 @@
 package com.ql.BlogApplication.service;
+
 import com.ql.BlogApplication.dto.ApiResponse;
 import com.ql.BlogApplication.dto.CategoryRequestDto;
 import com.ql.BlogApplication.entity.Category;
+import com.ql.BlogApplication.entity.Post;
 import com.ql.BlogApplication.repository.CategoryRepository;
+import com.ql.BlogApplication.repository.PostRepository;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -16,6 +19,7 @@ import java.util.Optional;
 public class CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final PostRepository postRepository;
 
     //returns all category list from the category table
     public ResponseEntity<ApiResponse<List<Category>>> getAllCategory(){
@@ -40,16 +44,31 @@ public class CategoryService {
            return new ResponseEntity<>(apiResponse,HttpStatus.OK);
     }
 
-    //delete a category and all posts under this category
+    //delete a category and assign all post under it as uncategorized
     public ResponseEntity<ApiResponse<String>> deleteCategory(Long id) {
         Optional<Category> optionalCategory=categoryRepository.findById(id);
+        Optional<Category> optionalUncategorized=categoryRepository.findByName("uncategorized");
 
         if(optionalCategory.isEmpty()){
             ApiResponse<String> apiResponse= ApiResponse.error(HttpStatus.NOT_FOUND.value(),"Category not found","Category not found");
             return new ResponseEntity<>(apiResponse,HttpStatus.NOT_FOUND);
         }
 
+        if(optionalUncategorized.isEmpty()){
+            ApiResponse<String> apiResponse= ApiResponse.error(HttpStatus.NOT_FOUND.value(),"Category not found","Category not found");
+            return new ResponseEntity<>(apiResponse,HttpStatus.NOT_FOUND);
+        }
+
+
+        List<Post> posts=optionalCategory.get().getPosts();
+
+        for(Post post:posts){
+              post.setCategory(optionalUncategorized.get());
+              postRepository.save(post);
+        }
+
         categoryRepository.deleteById(id);
+
         ApiResponse<String> apiResponse= ApiResponse.success(HttpStatus.OK.value(),"Category deleted successfully","Category deleted successfully");
         return new ResponseEntity<>(apiResponse,HttpStatus.OK);
     }
